Move types down to Workspaces
