+++ src/Features/Core/Portable/xlf/FeaturesResources.pl.xlf
@@ -1914,11 +1914,6 @@ If the "g" format specifier is used without other custom format specifiers, it's
         <target state="translated">Symbol nie ma ikony.</target>
         <note />
       </trans-unit>
-      <trans-unit id="Unknown">
-        <source>Unknown</source>
-        <target state="translated">Nieznany</target>
-        <note />
-      </trans-unit>
       <trans-unit id="Asynchronous_method_cannot_have_ref_out_parameters_colon_bracket_0_bracket">
         <source>Asynchronous method cannot have ref/out parameters : [{0}]</source>
         <target state="translated">Metoda asynchroniczna nie może zawierać parametrów ref/out: [{0}]</target>
