+++ src/Features/Core/Portable/xlf/FeaturesResources.ru.xlf
@@ -1914,11 +1914,6 @@ If the "g" format specifier is used without other custom format specifiers, it's
         <target state="translated">Символ не имеет значка.</target>
         <note />
       </trans-unit>
-      <trans-unit id="Unknown">
-        <source>Unknown</source>
-        <target state="translated">Unbekannt</target>
-        <note />
-      </trans-unit>
       <trans-unit id="Asynchronous_method_cannot_have_ref_out_parameters_colon_bracket_0_bracket">
         <source>Asynchronous method cannot have ref/out parameters : [{0}]</source>
         <target state="translated">Асинхронный метод не может иметь параметры ref/out: [{0}]</target>
