+++ src/Workspaces/Core/Portable/Diagnostics/DocumentDiagnosticAnalyzer.cs
