+++ src/Workspaces/Core/Portable/Diagnostics/FileContentLoadAnalyzer.cs
