+++ src/Workspaces/Core/Portable/Diagnostics/ProjectDiagnosticAnalyzer.cs
