+++ src/Workspaces/Core/Portable/WorkspacesResources.resx
@@ -1329,4 +1329,7 @@ Zero-width positive lookbehind assertions are typically used at the beginning of
   <data name="Solution_does_not_contain_specified_reference" xml:space="preserve">
     <value>Solution does not contain specified reference</value>
   </data>
+  <data name="Unknown" xml:space="preserve">
+    <value>Unknown</value>
+  </data>
 </root>
