+++ src/Workspaces/Core/Portable/xlf/WorkspacesResources.fr.xlf
@@ -1297,6 +1297,11 @@ Les assertions arrière positives de largeur nulle sont généralement utilisée
         <target state="new">The solution already contains the specified reference.</target>
         <note />
       </trans-unit>
+      <trans-unit id="Unknown">
+        <source>Unknown</source>
+        <target state="new">Unknown</target>
+        <note />
+      </trans-unit>
       <trans-unit id="Visual_Basic_files">
         <source>Visual Basic files</source>
         <target state="translated">Fichiers Visual Basic</target>
