+++ src/Workspaces/Core/Portable/xlf/WorkspacesResources.it.xlf
@@ -1297,6 +1297,11 @@ Le asserzioni lookbehind positive di larghezza zero vengono usate in genere all'
         <target state="new">The solution already contains the specified reference.</target>
         <note />
       </trans-unit>
+      <trans-unit id="Unknown">
+        <source>Unknown</source>
+        <target state="new">Unknown</target>
+        <note />
+      </trans-unit>
       <trans-unit id="Visual_Basic_files">
         <source>Visual Basic files</source>
         <target state="translated">File Visual Basic</target>
