+++ src/Workspaces/Core/Portable/xlf/WorkspacesResources.pt-BR.xlf
@@ -1297,6 +1297,11 @@ As declarações de lookbehind positivas de largura zero normalmente são usadas
         <target state="new">The solution already contains the specified reference.</target>
         <note />
       </trans-unit>
+      <trans-unit id="Unknown">
+        <source>Unknown</source>
+        <target state="new">Unknown</target>
+        <note />
+      </trans-unit>
       <trans-unit id="Visual_Basic_files">
         <source>Visual Basic files</source>
         <target state="translated">Arquivos do Visual Basic</target>
